Move all environment variables utils in a single file
